refactor(shared): move internal/crypto → shared/crypto
