+++ cmd/agentdock/app.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/Ivantseng123/agentdock/internal/bot"
 	"github.com/Ivantseng123/agentdock/internal/config"
-	"github.com/Ivantseng123/agentdock/internal/crypto"
+	"github.com/Ivantseng123/agentdock/shared/crypto"
 	ghclient "github.com/Ivantseng123/agentdock/internal/github"
 	"github.com/Ivantseng123/agentdock/internal/logging"
 	"github.com/Ivantseng123/agentdock/internal/mantis"
