+++ internal/worker/executor.go
@@ -11,7 +11,7 @@ import (
 	"time"
 
 	"github.com/Ivantseng123/agentdock/internal/bot"
-	"github.com/Ivantseng123/agentdock/internal/crypto"
+	"github.com/Ivantseng123/agentdock/shared/crypto"
 	"github.com/Ivantseng123/agentdock/shared/queue"
 )
