+++ shared/crypto/beacon.go
